MLIBZ-3057, updated save batch items request logic, improved errors return for batch post and single put requests separately.

+++ java-api-core/src/com/kinvey/java/model/KinveyBatchInsertError.java
@@ -5,7 +5,7 @@ import com.google.api.client.util.Key;
 public class KinveyBatchInsertError {
 
     @Key
-    private long index;
+    private int index;
     @Key
     private long code;
     @Key
@@ -27,11 +27,11 @@ public class KinveyBatchInsertError {
         this.code = code;
     }
 
-    public long getIndex() {
+    public int getIndex() {
         return index;
     }
 
-    public void setIndex(long index) {
+    public void setIndex(int index) {
         this.index = index;
     }
 }
